+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Implemented GELF Logger for sending logs to Graylog servers with support for both UDP and TCP protocols
 - Added server configuration for rate limits, CORS, additional HTTP headers
 - Implemented compression for GELF (gzip, zlib, none) with unit tests
+- Implemented GELF message truncation based on `max_message_size` configuration.
 
 ### Changed
 -